Added InputParser for list of id-ranges

+++ src/main/java/aoc/day02/IDRange.java
@@ -1,3 +1,3 @@
 package aoc.day02;
 
-public record IDRange(int lowerBound, int upperBound) {}
+public record IDRange(long lowerBound, long upperBound) {}
